Refactoring, move mapping into its own module
